+++ p2panda-rs/src/schema/validate/error.rs
@@ -47,6 +47,14 @@ pub enum ValidationError {
     /// Error from conversion of PlainValues.
     #[error(transparent)]
     NotStringValue(#[from] PlainValueError),
+
+    /// Error from validating system schema: `blob_v1`.
+    #[error("invalid 'blob_v1' operation: {0}")]
+    InvalidBlob(#[from] BlobError),
+
+    /// Error from validating system schema: `blob_piece_v1`.
+    #[error("invalid 'blob_piece_v1' operation: {0}")]
+    InvalidBlobPiece(#[from] BlobPieceError),
 }
 
 /// Custom error types for validating operations against `schema_field_definition_v1` schema.
@@ -85,4 +93,30 @@ pub enum SchemaDefinitionError {
     /// Error from conversion of PlainValues.
     #[error(transparent)]
     NotStringValue(#[from] PlainValueError),
+}
+
+/// Custom error types for validating operations against `blob_v1` schema.
+#[derive(Error, Debug)]
+#[allow(missing_copy_implementations)]
+pub enum BlobError {
+    /// "length" is greater than the maximum allowed as per specification.
+    #[error("'length' field in blob is over maximum allowed length")]
+    LengthInvalid,
+
+    /// "mime_type" is not correctly formatted as per specification.
+    #[error("'mime_type' field in blob is wrongly formatted")]
+    MimeTypeInvalid,
+
+    /// "pieces" is not correctly formatted as per specification.
+    #[error("'pieces' field in blob is wrongly formatted")]
+    PiecesInvalid,
+}
+
+/// Custom error types for validating operations against `blob_piece_v1` schema.
+#[derive(Error, Debug)]
+#[allow(missing_copy_implementations)]
+pub enum BlobPieceError {
+    /// "data" is greater than the maximum allowed as per specification.
+    #[error("'data' field in blob is over maximum allowed length")]
+    DataInvalid,
 }