+++ p2panda-rs/src/schema/validate/fields.rs
@@ -12,7 +12,8 @@ use crate::operation::{
 };
 use crate::schema::validate::error::ValidationError;
 use crate::schema::validate::{
-    validate_schema_definition_v1_fields, validate_schema_field_definition_v1_fields,
+    validate_blob_piece_v1_fields, validate_blob_v1_fields, validate_schema_definition_v1_fields,
+    validate_schema_field_definition_v1_fields,
 };
 use crate::schema::{FieldName, FieldType, Schema, SchemaId};
 
@@ -321,6 +322,14 @@ fn validate_system_schema_fields(
 ) -> Result<(), ValidationError> {
     match schema.id() {
         SchemaId::Application(_, _) => Ok(()),
+        SchemaId::Blob(_) => {
+            validate_blob_v1_fields(fields)?;
+            Ok(())
+        }
+        SchemaId::BlobPiece(_) => {
+            validate_blob_piece_v1_fields(fields)?;
+            Ok(())
+        }
         SchemaId::SchemaDefinition(_) => {
             validate_schema_definition_v1_fields(fields)?;
             Ok(())
@@ -344,6 +353,7 @@ mod tests {
     use crate::test_utils::constants::{HASH, SCHEMA_ID};
     use crate::test_utils::fixtures::document_view_id;
     use crate::test_utils::fixtures::schema_id;
+    use crate::test_utils::generate_random_bytes;
 
     use super::{
         validate_all_fields, validate_field, validate_field_name, validate_field_value,
@@ -806,6 +816,22 @@ mod tests {
         ],
         "invalid 'schema_field_definition_v1' operation: 'type' field in schema field definitions is wrongly formatted"
     )]
+    #[case::invalid_blob_mime_type(
+        SchemaId::Blob(1),
+        vec![
+            ("length", 100.into()),
+            ("mime_type", "wrong/mime/type/format".into()),
+            ("pieces", PlainValue::PinnedRelationList(vec![vec![HASH.to_owned()]])),
+         ],
+         "invalid 'blob_v1' operation: 'mime_type' field in blob is wrongly formatted"
+    )]
+    #[case::invalid_blob_piece_data(
+        SchemaId::BlobPiece(1),
+        vec![
+            ("data", generate_random_bytes(512 * 1000).into()),
+        ],
+        "invalid 'blob_piece_v1' operation: 'data' field in blob is over maximum allowed length"
+    )]
     fn wrong_system_schema_operations(
         #[case] schema_id: SchemaId,
         #[case] fields: Vec<(&str, PlainValue)>,
@@ -846,6 +872,20 @@ mod tests {
             ("name", "is_cute".into()),
             ("type", "bool".into()),
         ],
+    )]
+    #[case(
+        SchemaId::Blob(1),
+        vec![
+            ("length", 1.into()),
+            ("mime_type", "image/png".into()),
+            ("pieces", PlainValue::PinnedRelationList(vec![vec![HASH.to_owned()]])),
+         ]
+    )]
+    #[case(
+        SchemaId::BlobPiece(1),
+        vec![
+            ("data", "aGVsbG8gbXkgbmFtZSBpcyBzYW0=".into()),
+        ]
     )]
     fn correct_system_schema_operations(
         #[case] schema_id: SchemaId,